+++ intelmq/tests/bots/experts/rfc1918/test_expert.py
@@ -32,8 +32,16 @@ INPUT_DOMAIN = {"__type": "Event",
 OUTPUT_DOMAIN = {"__type": "Event",
                  "time.observation": "2015-01-01T00:00:00+00:00",
                  }
-INPUT_URL = {"__type": "Event",
+INPUT_URL1 = {"__type": "Event",
              "source.url": "http://sub.example.com/foo/bar",
+             "time.observation": "2015-01-01T00:00:00+00:00",
+             }
+INPUT_URL2 = {"__type": "Event",
+             "source.url": "http://192.168.0.1/foo/bar",
+             "time.observation": "2015-01-01T00:00:00+00:00",
+             }
+INPUT_ASN = {"__type": "Event",
+             "source.asn": 64496,
              "time.observation": "2015-01-01T00:00:00+00:00",
              }
 
@@ -47,8 +55,8 @@ class TestRFC1918ExpertBot(test.BotTestCase, unittest.TestCase):
     def set_bot(cls):
         cls.bot_reference = RFC1918ExpertBot
         cls.sysconfig = {'fields': 'destination.ip,source.ip,source.fqdn,'
-                                   'destination.fqdn,source.url',
-                         'policy': 'del,drop,drop,del,drop',
+                                   'destination.fqdn,source.url,source.url,source.asn',
+                         'policy': 'del,drop,drop,del,drop,drop,drop',
                          }
 
     def test_del(self):
@@ -71,8 +79,18 @@ class TestRFC1918ExpertBot(test.BotTestCase, unittest.TestCase):
         self.run_bot()
         self.assertMessageEqual(0, OUTPUT_DOMAIN)
 
-    def test_drop_url(self):
-        self.input_message = INPUT_URL
+    def test_drop_url1(self):
+        self.input_message = INPUT_URL1
+        self.run_bot()
+        self.assertOutputQueueLen(0)
+
+    def test_drop_url2(self):
+        self.input_message = INPUT_URL2
+        self.run_bot()
+        self.assertOutputQueueLen(0)
+
+    def test_drop_asn(self):
+        self.input_message = INPUT_ASN
         self.run_bot()
         self.assertOutputQueueLen(0)